Remove cricket::MediaType as a separate enum definition
Using constants and alias names to ensure that existing code compiles.
The part that is not backwards compatible is switch statements that
don't have default; they will react to the new value ANY.

Bug: webrtc:42222911
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/381660
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44164}

+++ sdk/android/src/jni/pc/media_stream_track.cc
@@ -21,7 +21,7 @@ namespace jni {
 ScopedJavaLocalRef<jobject> NativeToJavaMediaType(
     JNIEnv* jni,
     cricket::MediaType media_type) {
-  return Java_MediaType_fromNativeIndex(jni, media_type);
+  return Java_MediaType_fromNativeIndex(jni, static_cast<int>(media_type));
 }
 
 cricket::MediaType JavaToNativeMediaType(JNIEnv* jni,
